@@ -4,86 +4,87 @@ publish_to: 'none'
 version: 0.4.1+1
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.5.1 <4.0.0'
 
 dependencies:
   clock: ^1.1.1
   collection: ^1.18.0
-  cupertino_icons: ^1.0.6
-  device_info_plus: ^9.1.0
-  drift: ^2.13.1
+  cupertino_icons: ^1.0.8
+  device_info_plus: ^10.1.2
+  drift: ^2.20.0
   equatable: ^2.0.5
-  faker: ^2.1.0
-  file_picker: ^6.1.1
-  fl_chart: ^0.64.0
+  faker: ^2.2.0
+  file_picker: ^8.1.2
+  fl_chart: ^0.69.0
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-  flutter_riverpod: ^2.4.6
-  flutter_secure_storage: ^9.0.0
-  flutter_slidable: ^3.0.1
+  flutter_riverpod: ^2.5.1
+  flutter_secure_storage: ^9.2.2
+  flutter_slidable: ^3.1.1
   flutter_sticky_header: ^0.6.5
-  intl: ^0.18.1
+  intl: ^0.19.0
   logging: ^1.2.0
-  meta: ^1.11.0
+  meta: ^1.15.0
   money:
     git: https://github.com/jogboms/money.dart.git
-  package_info_plus: ^4.2.0
-  path: ^1.8.3
-  path_provider: ^2.1.1
-  platform: ^3.1.3
+  package_info_plus: ^8.0.2
+  path: ^1.9.0
+  path_provider: ^2.1.4
+  platform: ^3.1.5
   registry:
     git: https://github.com/jogboms/registry.dart.git
-  riverpod: ^2.4.6
-  riverpod_annotation: ^2.3.1
-  rxdart: ^0.27.7
-  share_plus: ^7.2.1
-  shared_preferences: ^2.2.2
+  riverpod: ^2.5.1
+  riverpod_annotation: ^2.3.5
+  rxdart: ">=0.27.0 <0.28.0"
+  share_plus: ^10.0.2
+  shared_preferences: ^2.3.2
   sliver_tools: ^0.2.12
-  sqlite3_flutter_libs: ^0.5.18
+  sqlite3_flutter_libs: ^0.5.24
   tabler_icons: ^2.5.0+1
   universal_io: ^2.2.2
-  url_launcher: ^6.2.1
-  uuid: ^4.2.1
+  url_launcher: ^6.3.0
+  uuid: ^4.5.0
 
 dependency_overrides:
-  meta: 1.11.0 # required by analyzer
+  # Because of https://github.com/invertase/dart_custom_lint/issues/228
+  custom_lint: 0.6.4
 
 dev_dependencies:
   args:
-  build_runner: ^2.4.6
-  device_info_plus_platform_interface: ^7.0.0
-  dotenv:
-  drift_dev: ^2.13.2
+  build_runner: ^2.4.12
+  custom_lint:
+  device_info_plus_platform_interface: ^7.0.1
+  drift_dev: ^2.20.1
   fake_async: ^1.3.1
   flutter_launcher_icons: ^0.13.1
   flutter_test:
     sdk: flutter
   mfsao: ^3.0.0
-  mocktail: ^1.0.1
-  package_info_plus_platform_interface: ^2.0.1
-  plugin_platform_interface: ^2.1.7
-  riverpod_generator: ^2.3.6
-  riverpod_lint: ^2.3.4
-  test: ^1.24.9
+  mocktail: ^1.0.4
+  package_info_plus_platform_interface: ^3.0.1
+  plugin_platform_interface: ^2.1.8
+  riverpod_generator:
+  riverpod_lint:
+  test:
 
 flutter:
   uses-material-design: true
   generate: true
 
   fonts:
-    - family: Poppins
+    - family: Inter
       fonts:
-        - asset: assets/fonts/Poppins-Thin.ttf
+        - asset: assets/fonts/Inter-Thin.ttf
           weight: 100
-        - asset: assets/fonts/Poppins-Light.ttf
+        - asset: assets/fonts/Inter-Light.ttf
           weight: 300
-        - asset: assets/fonts/Poppins-Regular.ttf
+        - asset: assets/fonts/Inter-Regular.ttf
           weight: 400
-        - asset: assets/fonts/Poppins-Medium.ttf
+        - asset: assets/fonts/Inter-Medium.ttf
           weight: 500
-        - asset: assets/fonts/Poppins-SemiBold.ttf
+        - asset: assets/fonts/Inter-SemiBold.ttf
           weight: 600
 
 flutter_launcher_icons:
