@@ -2,5 +2,4 @@ package io.github.jogboms.ovavue
 
 import io.flutter.embedding.android.FlutterActivity
 
-class MainActivity: FlutterActivity() {
-}
+class MainActivity: FlutterActivity()
