@@ -1,4 +1,4 @@
-// ignore_for_file: depend_on_referenced_packages
+#!/usr/bin/env dart
 
 import 'dart:io';
 
@@ -9,9 +9,9 @@ import 'utils.dart';
 final ArgParser parser = ArgParser()
   ..addOption(
     'versionBumpType',
-    allowed: VersionBumpType.names,
-    defaultsTo: VersionBumpType.patch.name,
-    valueHelp: VersionBumpType.patch.name,
+    allowed: VersionBumpType.values.names,
+    defaultsTo: VersionBumpType.build.name,
+    valueHelp: VersionBumpType.build.name,
   );
 
 void main(List<String> arguments) async {
@@ -19,6 +19,9 @@ void main(List<String> arguments) async {
   final String versionBumpType = args['versionBumpType'] as String;
 
   stdout.writeln(
-    await bumpPackageVersion(VersionBumpType.fromName(versionBumpType)),
+    await bumpPackageVersion(
+      VersionBumpType.fromName(versionBumpType),
+      workingDirectory: workingDirectory,
+    ),
   );
 }
